+++ extras/beast/test/string_istream.hpp
@@ -76,8 +76,8 @@ public:
     }
 
     template<class MutableBufferSequence, class ReadHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        ReadHandler, void(error_code, std::size_t))
+    async_return_type<
+        ReadHandler, void(error_code, std::size_t)>
     async_read_some(MutableBufferSequence const& buffers,
         ReadHandler&& handler)
     {
@@ -115,8 +115,8 @@ public:
     }
 
     template<class ConstBuffeSequence, class WriteHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        WriteHandler, void(error_code, std::size_t))
+    async_return_type<
+        WriteHandler, void(error_code, std::size_t)>
     async_write_some(ConstBuffeSequence const& buffers,
         WriteHandler&& handler)
     {