+++ include/beast/websocket/stream.hpp
@@ -970,8 +970,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept(AcceptHandler&& handler);
 
     /** Start reading and responding to a WebSocket HTTP Upgrade request.
@@ -1025,8 +1025,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class ResponseDecorator, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept_ex(ResponseDecorator const& decorator,
         AcceptHandler&& handler);
 
@@ -1079,8 +1079,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class ConstBufferSequence, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept(ConstBufferSequence const& buffers,
         AcceptHandler&& handler);
 
@@ -1143,8 +1143,8 @@ public:
     */
     template<class ConstBufferSequence,
         class ResponseDecorator, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept_ex(ConstBufferSequence const& buffers,
         ResponseDecorator const& decorator,
             AcceptHandler&& handler);
@@ -1195,8 +1195,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class Fields, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept(http::header<true, Fields> const& req,
         AcceptHandler&& handler);
 
@@ -1256,8 +1256,8 @@ public:
     */
     template<class Fields,
         class ResponseDecorator, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept_ex(http::header<true, Fields> const& req,
         ResponseDecorator const& decorator,
             AcceptHandler&& handler);
@@ -1316,8 +1316,8 @@ public:
     */
     template<class Fields,
         class ConstBufferSequence, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept(http::header<true, Fields> const& req,
         ConstBufferSequence const& buffers,
             AcceptHandler&& handler);
@@ -1385,8 +1385,8 @@ public:
     */
     template<class Fields, class ConstBufferSequence,
         class ResponseDecorator, class AcceptHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        AcceptHandler, void(error_code))
+    async_return_type<
+        AcceptHandler, void(error_code)>
     async_accept_ex(http::header<true, Fields> const& req,
         ConstBufferSequence const& buffers,
             ResponseDecorator const& decorator,
@@ -1863,8 +1863,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class HandshakeHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        HandshakeHandler, void(error_code))
+    async_return_type<
+        HandshakeHandler, void(error_code)>
     async_handshake(string_view const& host,
         string_view const& target,
             HandshakeHandler&& handler);
@@ -1915,8 +1915,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class HandshakeHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        HandshakeHandler, void(error_code))
+    async_return_type<
+        HandshakeHandler, void(error_code)>
     async_handshake(response_type& res,
         string_view const& host,
             string_view const& target,
@@ -1973,8 +1973,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class RequestDecorator, class HandshakeHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        HandshakeHandler, void(error_code))
+    async_return_type<
+        HandshakeHandler, void(error_code)>
     async_handshake_ex(string_view const& host,
         string_view const& target,
             RequestDecorator const& decorator,
@@ -2035,8 +2035,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class RequestDecorator, class HandshakeHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        HandshakeHandler, void(error_code))
+    async_return_type<
+        HandshakeHandler, void(error_code)>
     async_handshake_ex(response_type& res,
         string_view const& host,
             string_view const& target,
@@ -2145,8 +2145,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class CloseHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        CloseHandler, void(error_code))
+    async_return_type<
+        CloseHandler, void(error_code)>
     async_close(close_reason const& cr, CloseHandler&& handler);
 
     /** Send a WebSocket ping frame.
@@ -2223,8 +2223,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class WriteHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        WriteHandler, void(error_code))
+    async_return_type<
+        WriteHandler, void(error_code)>
     async_ping(ping_data const& payload, WriteHandler&& handler);
 
     /** Send a WebSocket pong frame.
@@ -2316,8 +2316,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class WriteHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        WriteHandler, void(error_code))
+    async_return_type<
+        WriteHandler, void(error_code)>
     async_pong(ping_data const& payload, WriteHandler&& handler);
 
     /** Read a message from the stream.
@@ -2460,8 +2460,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class DynamicBuffer, class ReadHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        ReadHandler, void(error_code))
+    async_return_type<
+        ReadHandler, void(error_code)>
     async_read(opcode& op, DynamicBuffer& dynabuf, ReadHandler&& handler);
 
     /** Read a message frame from the stream.
@@ -2614,8 +2614,8 @@ public:
         manner equivalent to using boost::asio::io_service::post().
     */
     template<class DynamicBuffer, class ReadHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        ReadHandler, void(error_code))
+    async_return_type<
+        ReadHandler, void(error_code)>
     async_read_frame(frame_info& fi,
         DynamicBuffer& dynabuf, ReadHandler&& handler);
 
@@ -2736,8 +2736,8 @@ public:
         manner equivalent to using `boost::asio::io_service::post`.
     */
     template<class ConstBufferSequence, class WriteHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        WriteHandler, void(error_code))
+    async_return_type<
+        WriteHandler, void(error_code)>
     async_write(ConstBufferSequence const& buffers,
         WriteHandler&& handler);
 
@@ -2848,8 +2848,8 @@ public:
         ); @endcode
     */
     template<class ConstBufferSequence, class WriteHandler>
-    BEAST_INITFN_RESULT_TYPE(
-        WriteHandler, void(error_code))
+    async_return_type<
+        WriteHandler, void(error_code)>
     async_write_frame(bool fin,
         ConstBufferSequence const& buffers, WriteHandler&& handler);